Open collapsible after custom error validation
Our custom error validation was getting in the way of some fields
being able to open the collapsible to focus the field in error.

This fixes that by stopping validation in an error case and having
the edit course form notice.

Stopping validation also prevents the submit button from spinning
forever.

DISCO-972
DISCO-973