Start Date bug
Users can change their course run start dates to
today or later, unless their start date has already
passed. In the latter case, they can change the start
date to the initially set start date or later (this may
result in changing a start date to a date in the past,
which is a small use case).

DISCO-1135