refactor: remove unused styles for autosuggest input
Removing these styles as part of the effor of reducing the use of
bootstrap variables in SCSS. Were able to remove these styles
completely, because the `.react-autosuggest__input` class is not applied
to the input element. Instead this element has class `form-control`,
because it's overwritten by `inputProps` in
`src/components/ListField/index.jsx` on lines 183 and 236.

In other words, these styles are never used, so removing them.